Create structs for packets; Cleanup module structure
And let the message-pack writer eat string slices instead of Strings.
Each packet now has a struct (in its own file).
The module tree is now in the 'lib.rs' file and not spread all over the place.
The imports of the 'main.rs' are a bit not so tasty right now due to this.