Better packet parsing and other small things
- CustomIterator now can return an empty vector when reading 0 bytes.
- MP-Writer: Added array writing.
- Added error handling at several places (LG-Formatter, ServerInstance).
- MessageType can now be checked for system and be converted to ID.
- Added 'rand' crate, to generate a random ID representing the server's network framework.
- The Lidgren part, now parses its extra data, instead of doing this in app scope.
- Formatting here and there...

The Lidgren packet parsing has been improved, in a sense that it not can create client objects which sort the packets in case of issues and defragmentation code has been added to each client.
The defragmentation needs a heartbeat to be performed - currently done before reading a packet.
Ping packets are now fully handled by the server code.