Improve error handling and add primitive logging
The new error handling now uses an object with a list of read-formatted messages. If all are printed, there is a stacktrace.
A type alias for the new Result exists, and it can be wrapped, to add more layers on the stack.

A logging framework has been started. Basically only introduces log levels which have different formatting colors.
The good thing about it is, that arguments get a different color.
Has yet to be applied onto the project.

Other:
- Improved many error messages
- Replaced panic with new exceptions
- Used some new log macros in the main file