Improve MessagePack reader
The previous system was written on demand basis.
Most reader functions did expect the type-prefix to be already read.
Which makes it impossible to use them on their own.

In the new system each type can be queried on its own.
And several functions to read Rust basic types - which automatically use whatever MP type is fit/available.

Optional has been removed from the type parsers.
It is not a macro that can be used on demand.

So now, one has to specify which Rust type is expected and will get exactly that.