Parse provided world file; Git-ignore data folder
Added a 'data' folder - it is relative to the execution path.
Will be ignored for git.
Inside is a whole world under the 'World' folder.
Only the world data file will be parsed for now.
But it is fully parsed. As next step the parsed data will be somehow saved and sent to the client.