Fix and enhance MP reader/writer
The MP reader was fully derped, as it had all 0x7F constants on 0x79.
Luckily noticed this soon enough.
The MP writer had a missing type byte when writing 8 bit ints.

Writing map and array is now implemented for all ranges.

Now there is a method to write i32 and bytes.