Allow CustomIterator to own data
Up to this point, CustomIterator could only borrow data, which had a few followup lifetime issues.
It now may contain the data, allowing to carrying it around in its instance.

Internally the code used Index[] accessors. Feature could not be ported to the new internal data structures, as lifetimes cannot be created on structs without lifetime-fields.