Fix and improve packet decompression
Previous commit broke it, as it did not stripe the packet ID off the bytes.
This commit fixes that, and it also reuses the already created Iterator.
The advantage of the new code is, that Iterators can now own data, hence it can easily be passed into an upper scope.
And its more pretty this way.