Fix macro unwrap_or_return
The macro assumed arguments for Options and closures for Results.
Closures allow side effects.

However, it was not possible to supply a simple argument for Result, nor a side-effect (via closure) for Options.
When supplying a block as argument, it would execute that regardless, when parsing it as function argument.

The only valid fix is to do what the Rust library does everywhere: Using two macros, one for arguments one for closures.